requirements.txt
fastapi
uvicorn
sqlalchemy
pymupdf
python-multipart
pillow
numpy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt] 